separate into SimpleFullbodyInverseKinematicsSolver.h
